+++ modules.json
@@ -86,5 +86,38 @@
 		"url": "https://github.com/totaljs/modules/tree/master/Miscellaneous/directorylisting",
 		"datecreated": "2016-01-01T00:00:00.000Z",
 		"dateupdated": null
+	},
+	{
+		"category": "Miscellaneous",
+		"name": "Monitor",
+		"version": "1.3.0",
+		"total": "2.0.0",
+		"description": "This module can monitor your Total.js web application and you can use AppMonitor with it.",
+		"author": "Peter Širka",
+		"url": "https://github.com/totaljs/modules/tree/master/Miscellaneous/monitor",
+		"datecreated": "2016-01-01T00:00:00.000Z",
+		"dateupdated": null
+	},
+	{
+		"category": "Miscellaneous",
+		"name": "Request stats",
+		"version": "1.1.0",
+		"total": "2.0.0",
+		"description": "This module counts all requests (dynamic content + files) by months. Data are stored in JSON file.",
+		"author": "Peter Širka",
+		"url": "https://github.com/totaljs/modules/tree/master/Miscellaneous/reqstats",
+		"datecreated": "2016-01-01T00:00:00.000Z",
+		"dateupdated": null
+	},
+	{
+		"category": "Miscellaneous",
+		"name": "Session",
+		"version": "2.0.0",
+		"total": "2.0.0",
+		"description": "This module enables session for each visitor. Session identificator is stored in cookie.",
+		"author": "Peter Širka",
+		"url": "https://github.com/totaljs/modules/tree/master/Miscellaneous/session",
+		"datecreated": "2016-01-01T00:00:00.000Z",
+		"dateupdated": null
 	}
 ]
